# docs/v1/swagger.yaml
basePath: /api/web-api-gateway/v1
definitions:
...
  domain.ErrorResponse:
    properties:
      code:
        type: integer
      data: {}
      is_error:
        type: boolean
      message:
        type: string
    type: object
...
  domain.SectionCreate:
    properties:
      name:
        example: purchase_planning_access
        type: string
    required:
    - name
    type: object
  domain.SectionUpdate:
    properties:
      name:
        example: status_and_calculation_access
        type: string
    required:
    - name
    type: object
...
  domain.SuccessResponse:
    properties:
      data: {}
      is_error:
        type: boolean
      total_count:
        type: integer
    type: object
...
host: 91.243.71.100:8080
info:
  contact:
    email: ru.system.ru@gmail.com
    name: ru.system.ru@gmail.com
  description: API gateway
  title: Web api gateway API
  version: "1.0"
paths:
...
  /sections:
    get:
      consumes:
      - application/json
      description: Получение списка секций
      operationId: get-sections
      parameters:
      - description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort
        required: true
        type: string
      - default: name
        description: Field to sort by
        enum:
        - id
        - name
        in: query
        name: sort_field
        required: true
        type: string
      - description: limit query param
        in: query
        name: limit
        required: true
        type: integer
      - description: offset query param
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get sections
      tags:
      - sections
    post:
      consumes:
      - application/json
      description: |-
        Создание секции
        Только super admin может создавать секции
      operationId: create-section
      parameters:
      - description: Section
        in: body
        name: section
        required: true
        schema:
          $ref: '#/definitions/domain.SectionCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create section
      tags:
      - sections
  /sections/{id}:
    delete:
      consumes:
      - application/json
      description: Удаление секции
      operationId: delete-section
      parameters:
      - description: Section ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete section
      tags:
      - sections
    get:
      consumes:
      - application/json
      description: Получение секции по id
      operationId: get-section
      parameters:
      - description: Section ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get section
      tags:
      - sections
    put:
      consumes:
      - application/json
      description: Обновление секции
      operationId: update-section
      parameters:
      - description: Section ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: Section
        in: body
        name: section
        required: true
        schema:
          $ref: '#/definitions/domain.SectionUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update section
      tags:
      - sections
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"